requirements.txt
openai==1.6.1
pytest==7.4.3
python-dotenv==1.0.0
